+++ R/r2_coxsnell.R
@@ -64,6 +64,10 @@ r2_coxsnell <- function(model) {
 
 #' @export
 r2_coxsnell.glm <- function(model) {
+  # if no deviance, return NA
+  if (is.null(model$deviance)) {
+    return(NULL)
+  }
   r2_coxsnell <- (1 - exp((model$deviance - model$null.deviance) / insight::n_obs(model)))
   names(r2_coxsnell) <- "Cox & Snell's R2"
   r2_coxsnell
@@ -122,6 +126,10 @@ r2_coxsnell.clm2 <- function(model) {
 #' @export
 r2_coxsnell.clm <- function(model) {
   l_base <- stats::logLik(stats::update(model, ~1))
+  # if no loglik, return NA
+  if (length(as.numeric(l_base)) == 0) {
+    return(NULL)
+  }
   .r2_coxsnell(model, l_base)
 }