+++ R/r2_nagelkerke.R
@@ -51,7 +51,11 @@ r2_nagelkerke <- function(model) {
 
 #' @export
 r2_nagelkerke.glm <- function(model) {
-  r2_nagelkerke <- r2_coxsnell(model) / (1 - exp(-model$null.deviance / insight::n_obs(model)))
+  r2cox <- r2_coxsnell(model)
+  if (is.na(r2cox) || is.null(r2cox)) {
+    return(NULL)
+  }
+  r2_nagelkerke <- r2cox / (1 - exp(-model$null.deviance / insight::n_obs(model)))
   names(r2_nagelkerke) <- "Nagelkerke's R2"
   r2_nagelkerke
 }
@@ -90,6 +94,10 @@ r2_nagelkerke.clm2 <- function(model) {
 #' @export
 r2_nagelkerke.clm <- function(model) {
   l_base <- stats::logLik(stats::update(model, ~1))
+  # if no loglik, return NA
+  if (length(as.numeric(l_base)) == 0) {
+    return(NULL)
+  }
   .r2_nagelkerke(model, l_base)
 }